requirements.txt
blinker==1.6.2
boto3==1.28.45
botocore==1.31.45
cffi==1.15.1
click==8.1.7
configparser==6.0.0
cryptography==41.0.3
dnspython==2.4.2
Flask-Cors==4.0.0
flask==2.3.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
pip==22.3.1
pycparser==2.21
PyJWT==2.8.0
pymongo==4.5.0
python-dateutil==2.8.2
s3transfer==0.6.2
setuptools==65.5.1
six==1.16.0
typing-extensions==4.8.0
urllib3==1.26.16
werkzeug==2.3.7
wheel==0.38.4
